Load content information from server in product page part 3

+++ app/src/main/java/com/example/digikala/data/models/product/ProductPageData.kt
@@ -11,7 +11,7 @@ data class Result(
 )
 
 data class Product(
-    val id: Double,
+    val id: Long,
     val category_id: Double,
     val category_title: String,
     val title_fa: String,
